Correct warning for possible data loss (conversion from const float to int).

+++ tutorial/detection/object/tutorial-detection-object-mbt.cpp
@@ -157,7 +157,7 @@ int main(int argc, char ** argv) {
     //! [Display reference keypoints]
     vpDisplay::display(I);
     for(std::vector<cv::KeyPoint>::const_iterator it = trainKeyPoints.begin(); it != trainKeyPoints.end(); ++it) {
-      vpDisplay::displayCross(I, it->pt.y, it->pt.x, 4, vpColor::red);
+      vpDisplay::displayCross(I, (int) it->pt.y, (int) it->pt.x, 4, vpColor::red);
     }
     vpDisplay::displayText(I, 10, 10, "Learning step: keypoints are detected on visible teabox faces", vpColor::red);
     vpDisplay::displayText(I, 30, 10, "Click to continue with detection...", vpColor::red);
